Changing demo controller time variable to double type
git-svn-id: https://atrias.googlecode.com/svn/trunk@1411 a504bc44-2baa-4cb8-4136-ba805e94d278

+++ drl-sim/atrias_controllers/ac_demo/include/ac_demo/controller_core.h
@@ -36,7 +36,7 @@
 
 uint8_t currentState;
 uint8_t currentDemo;
-float elapsedTime;
+double elapsedTime;
 RobotPosition desiredPos;
 RobotPosition lastDemoPos;
 float desTorqueA;
